Turn the project into a crate
