+++ Cargo.toml
@@ -3,6 +3,14 @@ name = "client-backend"
 version = "0.1.0"
 edition = "2021"
 
+[[bin]]
+name = "client_backend"
+path = "src/main.rs"
+
+[lib]
+name = "client_backend"
+path = "src/lib.rs"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
